% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_rearrange.R
\name{array_rearrange}
\alias{array_rearrange}
\title{rearrange array from n-d to m-d, with m < n}
\usage{
array_rearrange(original_array, flatten = c(1), position = min(flatten))
}
\arguments{
\item{original_array}{array to rearranged}

\item{flatten}{dimensions to be flattened.}

\item{position}{place /order/position of the new dimension}
}
\description{
array_rearrange (original_array=X,   flatten=c(2,3), position=1)
}
