% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brt_pre.R
\name{Brt_pre}
\alias{Brt_pre}
\title{brt predictions}
\usage{
Brt_pre(
  variabledf,
  opti = F,
  ntree = 1000,
  y_varname = c("day_value", "night_value", "value_mean"),
  interaction.depth = 6,
  bag.fraction = 0.5,
  shrinkage = 0.01,
  training,
  test,
  grepstring,
  ...
)
}
\description{
brt predictions
}
