% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_center.R
\name{crop_center}
\alias{crop_center}
\title{crop to center and quick view with mapview}
\usage{
crop_center(ras, size, vis = F, ...)
}
\arguments{
\item{ras}{raster or rasterstack}

\item{size}{windowsize for cropping}

\item{vis}{if True then apply mapview}
}
\description{
crop to center and quick view with mapview
}
