% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_h2o_pred.R
\name{get_h2o_pred}
\alias{get_h2o_pred}
\title{get h2o cv prediction}
\usage{
get_h2o_pred(h2omodel)
}
\arguments{
\item{h2omodel}{h2omodel}
}
\value{
k-fold cv prediction
}
\description{
get h2o cv prediction
}
