% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lasso_pre.R
\name{Lasso_pre}
\alias{Lasso_pre}
\title{Lasso predictions}
\usage{
Lasso_pre(
  variabledf,
  alpha = 1,
  y_varname = c("day_value", "night_value", "value_mean"),
  training,
  test,
  grepstring = "ROAD|pop|temp|wind|Rsp|OMI|eleva|coast"
)
}
\description{
Lasso predictions
}
