% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeraster2file.R
\name{mergeraster2file}
\alias{mergeraster2file}
\title{Merge new predictor variables (rasters)}
\usage{
mergeraster2file(tomerge, rasterfile, lon_lat_name, varname)
}
\arguments{
\item{tomerge}{the dataframe to merge the raster files to}

\item{rasterfile}{the R raster object (raster package) to merge}

\item{lon_lat_name}{a vector of the names of longitude and latitude. Commonly Longitude first and latitude second.}

\item{varname}{the name of the new variable}
}
\value{
a dataframe with merged variables
}
\description{
Merge new predictor variables (rasters)
}
\examples{
\donttest{merged = mergeraster2file(merged, c("lon","lat"), 'elevation.map', 'elevation')}
}
