% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicLA_RF_XGBtiles.R
\name{predicLA_RF_XGBtiles}
\alias{predicLA_RF_XGBtiles}
\title{predict tiles using LA, RF, SGB}
\usage{
predicLA_RF_XGBtiles(df, rasstack, yname, xgbname, rfname, laname,
  ntree = 1000, max_depth = 6, eta = 0.02, nthread = 4,
  nrounds = 1000, ...)
}
\arguments{
\item{df}{the dataframe for building the model}

\item{rasstack}{rasterstack, predictors}

\item{yname}{the y variable name}

\item{xgbname}{output filename for xgb}

\item{rfname}{output filename for rf}

\item{ntree}{RF ntree, default 1000}

\item{lanme}{output filename for LA}
}
\description{
predict tiles using LA, RF, SGB
}
\examples{
\donttest{
xgbname = "/data/lu01/NWA/xgb6-Jul_oaq.tif"
rfname = "/data/lu01/NWA/RF6-Juloaq.tif"
laname= "/data/lu01/NWA/LA6-Juloaq.tif"
lus = raster("/data/lu01/NWA/predictor/NLstack.grd")
lf_lo = list.files("/data/lu01/NWA/Bakfietsdata", pattern = "^.*morning.*.csv$", full.names = T)
bakfile1 = read.csv(lf_lo[1])
proj = "+proj=longlat +datum=WGS84"
df = retrieve_predictor(lus, bakfile1, c("Lon", "Lat"), proj)
wrapplot(df, lus, "NO2", xgbname=xgbname, rfname = rfname, laname = laname )}
}
