% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_with_pp_La.R
\name{prediction_with_pp_La}
\alias{prediction_with_pp_La}
\title{Aggregate random forest using lasso}
\usage{
prediction_with_pp_La(rfmodel, trainingXY, trainingY, testingXY)
}
\arguments{
\item{rfmodel}{ranger model}

\item{trainingXY}{training matrix, with same response and predictor names as in the rfmodel}

\item{trainingY}{training response vection}

\item{testingXY}{testing matrix, with same response and predictor names as in the rfmodel}
}
\description{
Aggregate random forest using lasso
}
