% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_rf_LUR.R
\name{q_rf_LUR}
\alias{q_rf_LUR}
\title{Quantile Random forest  (using ranger, quantregForest)}
\usage{
q_rf_LUR(
  variabledf,
  upq = "quantile= 0.05",
  lq = "quantile= 0.95",
  numtrees = 1000,
  mtry = NULL,
  y_varname = c("day_value", "night_value", "value_mean"),
  training,
  test,
  grepstring,
  ...
)
}
\arguments{
\item{variabledf}{the dataframe containing predictors and dependent variable}

\item{upq}{upper quantile default 0.95}

\item{lq}{lower quantile, default 0.05}

\item{y_varname}{name of the dependent variable.}

\item{training}{the index for the rows used for training.}

\item{test}{the index for the rows used for testing.}

\item{grepstring}{the variable/column names of predictors in Lasso, grepl stlye, e.g. 'ROAD|pop|temp|wind|Rsp|OMI|eleva|coast'}
}
\value{
an error matrix and coverage probability, default 0.90 prediction interval.
}
\description{
Quantile Random forest  (using ranger, quantregForest)
}
