% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_stack.R
\name{ranger_stack}
\alias{ranger_stack}
\title{apply ranger to rasterstrack}
\usage{
ranger_stack(df, yname, rasstack, mtry, num.trees)
}
\arguments{
\item{df}{dataframe with response and predictors}

\item{yname}{the response variable}

\item{rasstack}{the rasterstack for the model to be applied to}
}
\description{
apply ranger to rasterstrack
}
