% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removedips.R
\name{removedips}
\alias{removedips}
\title{remove dips}
\usage{
removedips(x, perce = 0.15)
}
\arguments{
\item{x}{a vector that the extreme values are to be removed}
}
\value{
time series with dips removed
}
\description{
remove time series points that is less than a certain percentage (by default 15\%) of the NO2 measured in the previous or the next time stamp
}
