% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_lasso_LUR.R
\name{rf_Lasso_LUR}
\alias{rf_Lasso_LUR}
\title{Accuracy matrix of aggregating random forest using lasso}
\usage{
rf_Lasso_LUR(
  variabledf,
  vis = F,
  numtrees = 2000,
  mtry = 33,
  y_varname = c("day_value", "night_value", "value_mean"),
  training,
  test,
  grepstring,
  ...
)
}
\arguments{
\item{variabledf}{the dataframe containing predictors and dependent variable}

\item{vis}{if true, plot variable importance}

\item{y_varname}{name of the dependent variable.}

\item{training}{the index for the rows used for training.}

\item{test}{the index for the rows used for testing.}

\item{grepstring}{the variable/column names of predictors in Lasso, grepl stlye, e.g. 'ROAD|pop|temp|wind|Rsp|OMI|eleva|coast'}
}
\value{
plot variable importance and an error matrix
}
\description{
Accuracy matrix of aggregating random forest using lasso
}
