% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_pre.R
\name{rf_pre}
\alias{rf_pre}
\title{rf predictions}
\usage{
rf_pre(
  variabledf,
  numtrees = 2000,
  mtry = 33,
  y_varname = c("day_value", "night_value", "value_mean"),
  test,
  training,
  grepstring = "ROAD|pop|temp|wind|Rsp|OMI|eleva|coast",
  ...
)
}
\description{
rf predictions
}
