% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_pre.R
\name{xgb_pre}
\alias{xgb_pre}
\title{xgb predictions}
\usage{
xgb_pre(
  variabledf,
  max_depth,
  eta,
  gamma,
  xgb_lambda,
  xgb_alpha,
  nrounds,
  subsample,
  y_varname = c("day_value", "night_value", "value_mean"),
  training,
  test,
  grepstring,
  ...
)
}
\description{
xgb predictions
}
