% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_stack.R
\name{xgb_stack}
\alias{xgb_stack}
\title{xgb predictions}
\usage{
xgb_stack(
  sr,
  df_var,
  y_var,
  xgbname = "xgb.tif",
  max_depth = 5,
  eta = 0.003,
  nthread = 32,
  gamma = 1,
  nrounds = 1650,
  verbose = 0,
  xgb_lambda = 1,
  xgb_alpha = 0.02
)
}
\arguments{
\item{sr}{raster stack of predictors}

\item{df_var}{dataframe containing predictors and responses}

\item{y_var}{specify the name of the response}

\item{xgbname}{specify the name/directoray of the predicted raster file to save}
}
\description{
XGboost prediction on rasterstacks, allowing missing value
}
\details{
use as.matrix instead of a sparse matrix model, the R Matrix package is therefore not needed and missing data is kept in, though in situations with lots of zeros, it is less efficient than using a sparse matrix. The same dense array is used in the xgboost_imp, xgboost_LUR, and xgb_pre. The other funciton, predicLA_RF_XGB_tiles used a sparse matrix, which means rows with missing values are missed.
}
