% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost_imp.R
\name{xgboost_imp}
\alias{xgboost_imp}
\title{return variable importance xgboost}
\usage{
xgboost_imp(
  variabledf,
  max_depth,
  eta,
  gamma,
  xgb_lambda,
  xgb_alpha,
  nrounds,
  subsample,
  y_varname = c("day_value", "night_value", "value_mean"),
  training = nrow(variabledf),
  grepstring,
  ...
)
}
\description{
return variable importance xgboost
}
