% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_pre.R
\name{xgb_pre}
\alias{xgb_pre}
\title{xgb predictions}
\usage{
xgb_pre(
  variabledf,
  max_depth = 4,
  eta = 0.02,
  nthread = 2,
  gamma = 0,
  nrounds = 300,
  subsample = 0.7,
  y_varname = c("day_value", "night_value", "value_mean"),
  training,
  test,
  grepstring,
  ...
)
}
\description{
xgb predictions
}
