% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost_imp.R
\name{xgboost_imp}
\alias{xgboost_imp}
\title{return variable importance xgboost}
\usage{
xgboost_imp(
  variabledf,
  max_depth = 4,
  eta = 0.02,
  gamma = 0,
  nthread = 2,
  nrounds = 300,
  subsample = 0.7,
  y_varname = c("day_value", "night_value", "value_mean"),
  training,
  test,
  grepstring,
  ...
)
}
\description{
return variable importance xgboost
}
